+++ core/execution/amm/engine_test.go
@@ -45,6 +45,7 @@ var (
 
 func TestSubmitAMM(t *testing.T) {
 	t.Run("test one pool per party", testOnePoolPerParty)
+	t.Run("test creation of sparse AMM", testSparseAMMEngine)
 }
 
 func TestAMMTrading(t *testing.T) {
@@ -154,7 +155,7 @@ func testAmendInsufficientCommitment(t *testing.T) {
 	amend.Parameters.LowerBound.AddSum(num.UintOne())
 
 	_, _, err := tst.engine.Amend(ctx, amend, riskFactors, scalingFactors, slippage)
-	require.ErrorContains(t, err, "insufficient commitment")
+	require.ErrorContains(t, err, "commitment amount too low")
 
 	// check that the original pool still exists
 	assert.Equal(t, poolID, tst.engine.poolsCpy[0].ID)
@@ -299,10 +300,10 @@ func testSubmitMarketOrder(t *testing.T) {
 	}
 
 	ensurePosition(t, tst.pos, 0, num.NewUint(0))
-	orders := tst.engine.SubmitOrder(agg, num.NewUint(1980), num.NewUint(1990))
+	orders := tst.engine.SubmitOrder(agg, num.NewUint(2000), num.NewUint(1980))
 	require.Len(t, orders, 1)
-	assert.Equal(t, "1994", orders[0].Price.String())
-	assert.Equal(t, 126420, int(orders[0].Size))
+	assert.Equal(t, "1989", orders[0].Price.String())
+	assert.Equal(t, 251890, int(orders[0].Size))
 }
 
 func testSubmitMarketOrderUnbounded(t *testing.T) {
@@ -488,7 +489,7 @@ func testBestPricesAndVolume(t *testing.T) {
 }
 
 func TestBestPricesAndVolumeNearBound(t *testing.T) {
-	tst := getTestEngineWithFactors(t, num.DecimalFromInt64(100), num.DecimalFromFloat(10))
+	tst := getTestEngineWithFactors(t, num.DecimalFromInt64(100), num.DecimalFromFloat(10), 0)
 
 	// create three pools
 	party, subAccount := getParty(t, tst)
@@ -497,7 +498,7 @@ func TestBestPricesAndVolumeNearBound(t *testing.T) {
 	expectSubaccountCreation(t, tst, party, subAccount)
 	whenAMMIsSubmitted(t, tst, submit)
 
-	tst.pos.EXPECT().GetPositionsByParty(gomock.Any()).Times(3).Return(
+	tst.pos.EXPECT().GetPositionsByParty(gomock.Any()).Times(7).Return(
 		[]events.MarketPosition{&marketPosition{size: 0, averageEntry: num.NewUint(0)}},
 	)
 
@@ -508,7 +509,7 @@ func TestBestPricesAndVolumeNearBound(t *testing.T) {
 	assert.Equal(t, 1192, int(avolume))
 
 	// lets move its position so that the fair price is within one tick of the AMMs upper boundary
-	tst.pos.EXPECT().GetPositionsByParty(gomock.Any()).Times(3).Return(
+	tst.pos.EXPECT().GetPositionsByParty(gomock.Any()).Times(7).Return(
 		[]events.MarketPosition{&marketPosition{size: -222000, averageEntry: num.NewUint(0)}},
 	)
 
@@ -516,17 +517,17 @@ func TestBestPricesAndVolumeNearBound(t *testing.T) {
 	assert.Equal(t, "219890", bid.String())
 	assert.Equal(t, "220000", ask.String()) // make sure we are capped to the boundary and not 220090
 	assert.Equal(t, 1034, int(bvolume))
-	assert.Equal(t, 103, int(avolume))
+	assert.Equal(t, 104, int(avolume))
 
 	// lets move its position so that the fair price is within one tick of the AMMs upper boundary
-	tst.pos.EXPECT().GetPositionsByParty(gomock.Any()).Times(3).Return(
+	tst.pos.EXPECT().GetPositionsByParty(gomock.Any()).Times(7).Return(
 		[]events.MarketPosition{&marketPosition{size: 270400, averageEntry: num.NewUint(0)}},
 	)
 
 	bid, bvolume, ask, avolume = tst.engine.BestPricesAndVolumes()
 	assert.Equal(t, "180000", bid.String()) // make sure we are capped to the boundary and not 179904
 	assert.Equal(t, "180104", ask.String())
-	assert.Equal(t, 58, int(bvolume))
+	assert.Equal(t, 62, int(bvolume))
 	assert.Equal(t, 1460, int(avolume))
 }
 
@@ -696,6 +697,26 @@ func testMarketClosure(t *testing.T) {
 	require.Equal(t, 0, len(tst.engine.ammParties))
 }
 
+func testSparseAMMEngine(t *testing.T) {
+	tst := getTestEngineWithFactors(t, num.DecimalOne(), num.DecimalOne(), 10)
+
+	party, subAccount := getParty(t, tst)
+	submit := getPoolSubmission(t, party, tst.marketID)
+	submit.CommitmentAmount = num.NewUint(100000)
+
+	expectSubaccountCreation(t, tst, party, subAccount)
+	whenAMMIsSubmitted(t, tst, submit)
+
+	tst.pos.EXPECT().GetPositionsByParty(gomock.Any()).AnyTimes().Return(
+		[]events.MarketPosition{&marketPosition{size: 0, averageEntry: nil}},
+	)
+	bb, bv, ba, av := tst.engine.BestPricesAndVolumes()
+	assert.Equal(t, "1992", bb.String())
+	assert.Equal(t, 1, int(bv))
+	assert.Equal(t, "2009", ba.String())
+	assert.Equal(t, 1, int(av))
+}
+
 func expectSubaccountCreation(t *testing.T, tst *tstEngine, party, subAccount string) {
 	t.Helper()
 
@@ -815,7 +836,7 @@ type tstEngine struct {
 	assetID  string
 }
 
-func getTestEngineWithFactors(t *testing.T, priceFactor, positionFactor num.Decimal) *tstEngine {
+func getTestEngineWithFactors(t *testing.T, priceFactor, positionFactor num.Decimal, allowedEmptyLevels uint64) *tstEngine {
 	t.Helper()
 	ctrl := gomock.NewController(t)
 	col := mocks.NewMockCollateral(ctrl)
@@ -836,7 +857,7 @@ func getTestEngineWithFactors(t *testing.T, priceFactor, positionFactor num.Deci
 	parties := cmocks.NewMockParties(ctrl)
 	parties.EXPECT().AssignDeriveKey(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
 
-	eng := New(logging.NewTestLogger(), broker, col, marketID, assetID, pos, priceFactor, positionFactor, mat, parties)
+	eng := New(logging.NewTestLogger(), broker, col, marketID, assetID, pos, priceFactor, positionFactor, mat, parties, allowedEmptyLevels)
 
 	// do an ontick to initialise the idgen
 	ctx := vgcontext.WithTraceID(context.Background(), vgcrypto.RandomHash())
@@ -856,7 +877,7 @@ func getTestEngineWithFactors(t *testing.T, priceFactor, positionFactor num.Deci
 
 func getTestEngine(t *testing.T) *tstEngine {
 	t.Helper()
-	return getTestEngineWithFactors(t, num.DecimalOne(), num.DecimalOne())
+	return getTestEngineWithFactors(t, num.DecimalOne(), num.DecimalOne(), 0)
 }
 
 func getAccount(balance uint64) *types.Account {